+++ types/ember-simple-auth/index.d.ts
@@ -1,1 +1,2 @@
 declare module 'ember-simple-auth/test-support';
+declare module 'ember-simple-auth/services/session';